daemon: move stream `jsonrpc_` methods to an individual module
